Mitigate SIGSEGV in BitmapFactory.decodeStream() on Android 4.4
On KitKat, occasional SIGSEGVs occur in BitmapFactory.decodeStream() under poor
network conditions. These crashes are preceded by a series of
SocketTimeoutExceptions and a single call to skip(). This appears to be a
platform bug specific to Android 4.4.

Additionally, when read() throws an IOException, it's repeatedly called in a
tight loop, throwing again each time.

This commit attempts to mitigate the issue by converting IOException in read()
to EOF on KitKat.