Fix installation fail on unsupported architectures
This is a dirty hack to avoid "Native ABI mismatch from package file"
error during the installation on devices with architectures that
Conscrypt doesn't provide a native .so library for. Unsupported
architectures' directories are populated with an empty
libconscrypt_jni.so file just to trick the installer and are not loaded
to avoid crashing with SIGBUS.

Another possible solution would be an external APK plugin that provides
Conscrypt library on demand. It would also keep the main APK small.
But as long as practical, I prefer to keep everything in a single
APK file.