Avoid restarting drawing thread
Ideally, the surface shouldn't be recreated when reordering CameraViews
in the grid, but this requires a full custom layout (which would be nice
to have for other reasons too). This is a quick workaround to keep the
drawing thread alive.

This change also shortens the UI thread blocking time in
surfaceDestroyed() as it now doesn't join() the drawing thread.