Fix drafter thread stuck forever in multi-threaded display
If a user quickly enters and exits a SingleCameraActivity,
MatrixActivity's onStop() may never be called, but the SurfaceViewDisplay's
drafter thread has already been abandoned. Because onStop() is not called,
CameraView is not recreated. CameraPlayer contiunes to decode but it
signals the drafter thread created in SingleCameraActivity
(SingleFrameBuffer.newConsumer() changes signaling target) even though the
activity has already finished and the thread itself is dead. This change
allows the drafter thread reclaim CameraPlayer's signaling.

There must be a better way to do this...